chore(component): refactoring
