+++ biproxi/package.json
@@ -7,7 +7,8 @@
     "build": "next build",
     "start": "next start",
     "lint": "next lint",
-    "make run": "node ssr-server.js"
+    "make run": "node ssr-server.js",
+    "make Test": "jest --watch './src/**/*.test.js'"
   },
   "dependencies": {
     "@reduxjs/toolkit": "^1.8.1",
@@ -22,15 +23,16 @@
     "styled-components": "^5.3.5"
   },
   "devDependencies": {
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^12.1.5",
     "@types/express": "^4.17.13",
     "@types/node": "17.0.25",
-    "@types/react": "^17.0.2",
-    "@types/react-dom": "^17.0.2",
-    "@types/react-modal": "^3.13.1",
-    "chai": "^4.3.6",
+    "@types/react": "^17.0.44",
+    "@types/react-dom": "^17.0.15",
     "eslint": "8.13.0",
     "eslint-config-next": "12.1.5",
-    "mocha": "^9.2.2",
+    "jest": "^27.5.1",
+    "jest-fetch-mock": "^3.0.3",
     "typescript": "4.6.3"
   }
 }
